TRAINER_TOOL/requirements_kaggle.txt
tqdm
numpy
opencv-python
pillow
pyyaml
torch
torchvision
timm
onnx
onnxscript
# optional
paddleocr
